Crack_Detection_with_FRCNN/requirements.txt
absl-py==0.11.0
astor==0.8.1
cached-property==1.5.2
certifi==2020.12.5
cycler==0.10.0
gast==0.2.2
google-pasta==0.2.0
grpcio==1.34.0
h5py==3.1.0
importlib-metadata==3.1.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.3
matplotlib==3.3.3
numpy==1.19.4
opt-einsum==3.3.0
pandas==1.1.5
Pillow==8.0.1
protobuf==3.14.0
pyparsing==2.4.7
PyQt5==5.14.2
python-dateutil==2.8.1
pytz==2020.4
scipy==1.5.4
seaborn==0.11.0
sip==4.19.8
six==1.15.0
tensorboard==1.15.0
tensorflow==1.15.2
tensorflow-estimator==1.15.1
termcolor==1.1.0
Werkzeug==1.0.1
wincertstore==0.2
wrapt==1.12.1
zipp==3.4.0
flask-Cors
flask-jsonpify
Flask-RESTful
requests